setup: initial version
This version does what it's 'supposed to do'. It exposes a serial device
on your system (usually `/dev/ttyACM0`) that you can send ASCII text to.
Though not all characters are supported.

It's fast enough to do some stuff; and it also has a very slight fade.
There's plenty of room for improvement of course, and improvement will
be coming.

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>