uart: change to pio
Import the example PIO UART [1] to get us started with using PIO instead
of the UART peripheral. This should (later on) allow us to more easily
use DMA and attach multiple displays.

For now we start with a sligthly adjusted example PIO; the adjustment is
that we don't send strings but explicit length.

[1]: https://github.com/raspberrypi/pico-examples/blob/master/pio/uart_tx/uart_tx.pio

Signed-off-by: Simon de Vlieger <supakeen@redhat.com>